ryan-tinystories/pyproject.toml
[project]
name = "ryan-tinystories"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.5.0",
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.2",
    "openai>=1.0.0",
    "peft>=0.18.1",
    "python-dotenv>=1.0.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
]
